pyproject.toml
[tool.poetry]
name = "DZevent_backend"
version = "0.1.0"
description = "Backend for events"
authors = ["Badis, Chakib, AbdelAli"]

[tool.poetry.dependencies]
python = ">=3.12.0,<3.13"
python-dotenv = "1.0.0"
django = "^6.0"
djangorestframework = "^3.16.1"
supabase = "^2.25.1"
postgres = "^4.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"